se corrige problema en id

+++ biblioteca-backend/src/main/java/com/libreria/biblioteca/service/IUsersService.java
@@ -3,6 +3,7 @@ package com.libreria.biblioteca.service;
 import com.libreria.biblioteca.models.Users;
 import org.springframework.stereotype.Service;
 
+import java.sql.SQLException;
 import java.util.List;
 import java.util.Optional;
 
@@ -12,11 +13,13 @@ public interface IUsersService {
 
     public void crearUsers(Users usuario);
 
-    public void bajarUsers(Integer id);
+    public void bajarUsers(Long id);
 
-    public Optional<Users> buscarUser(Integer id);
+    public Optional<Users> buscarUser(Long id);
 
     public void updateUser(Users u);
 
     public Optional<Users> buscarUserName(String nombre);
+
+  //  void InsertLending(int id, String bookid) throws SQLException;
 }
